backend/requirements.txt
# FastAPI backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.0

# Vector database
chromadb==0.4.18

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0

# ML and embeddings
openai>=1.12.0
numpy==1.26.2
scikit-learn==1.3.2

# For frontend (if running from same environment)
streamlit==1.29.0

